Syncup work on editor tooling

+++ Source/MyProject/Private/Items/ItemExamineWidget.cpp
@@ -8,18 +8,17 @@
 
 void UItemExamineWidget::NativeConstruct()
 {
-  btnClose->OnClicked.AddDynamic(this, &UItemExamineWidget::CloseMenu);
+   btnClose->OnClicked.AddDynamic(this, &UItemExamineWidget::CloseMenu);
 }
 
 bool UItemExamineWidget::OnWidgetAddToViewport_Implementation()
 {
-  if(itemToDisplayID != -1)
-     return true;
-  return false;
+   if(itemToDisplayID != -1) return true;
+   return false;
 }
 
 void UItemExamineWidget::CloseMenu()
 {
-  itemToDisplayID = 0;
-  hudManagerRef->AddHUD(static_cast<uint8>(EHUDs::HS_ExamineMenu));
+   itemToDisplayID = 0;
+   hudManagerRef->AddHUD(EHUDs::HS_ExamineMenu);
 }
